Extract common code into src.models.utils
